try to rm type of masses

+++ src/constants.jl
@@ -12,7 +12,7 @@ This value is taken from the PDG 2020 recommended values of the fundamental phys
 - The Review of Particle Physics: P.A. Zyla et al. (Particle Data Group), Prog. Theor. Exp. Phys. 2020, 083C01 (2020)
 
 """
-const ELECTRON_MASS::Float64 = 0.51099895000
+const ELECTRON_MASS = 0.51099895000
 
 """
    MUON_MASS 
@@ -25,7 +25,7 @@ This value is taken from the PDG 2020 recommended values of the fundamental phys
 - The Review of Particle Physics: P.A. Zyla et al. (Particle Data Group), Prog. Theor. Exp. Phys. 2020, 083C01 (2020)
 
 """
-const MUON_MASS::Float64 = 105.6583755
+const MUON_MASS = 105.6583755
 
 """
   ALPHA 
@@ -38,4 +38,4 @@ This value is taken from the PDG 2020 recommended values of the fundamental phys
 - The Review of Particle Physics: P.A. Zyla et al. (Particle Data Group), Prog. Theor. Exp. Phys. 2020, 083C01 (2020)
 
 """
-const ALPHA::Float64 = 0.0072973525
+const ALPHA = 0.0072973525
